# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Kultr Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: countryCode
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: dateFrom
          in: query
          description: Only include events on or after this ISO date/date-time.
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: Only include events on or before this ISO date/date-time.
          schema:
            type: string
            format: date
        - name: priceMin
          in: query
          description: >-
            Only include events with at least one ticket type priced at or
            above this amount.
          schema:
            type: number
        - name: priceMax
          in: query
          description: >-
            Only include events with at least one ticket type priced at or
            below this amount.
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
  /events/search:
    get:
      operationId: searchEvents
      tags: [events]
      summary: Search live events by free-text query, with optional filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
            default: ""
        - name: dateFrom
          in: query
          description: Only include events on or after this ISO date/date-time.
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: Only include events on or before this ISO date/date-time.
          schema:
            type: string
            format: date
        - name: priceMin
          in: query
          description: >-
            Only include events with at least one ticket type priced at or
            above this amount.
          schema:
            type: number
        - name: priceMax
          in: query
          description: >-
            Only include events with at least one ticket type priced at or
            below this amount.
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Matching events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
  /events/creator/me:
    get:
      operationId: listMyEvents
      tags: [events]
      summary: List events created by the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Creator events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
  /events/creator/analytics:
    get:
      operationId: getCreatorAnalytics
      tags: [events]
      summary: Get sales analytics for the creator's events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Creator analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorAnalytics"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/status:
    patch:
      operationId: updateEventStatus
      tags: [events]
      summary: >-
        Move an event through its lifecycle. A creator may submit their own
        draft for review (draft→pending_review), withdraw it
        (pending_review→draft), or cancel/end their own live event — they
        can never set an event live directly. Only an admin can approve a
        submission (pending_review→live), or force any transition on any
        event at any time (the moderation kill-switch).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStatusResponse"
        "403":
          description: Not the event's creator and not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: This transition is not allowed for a non-admin caller from the event's current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/admin/all:
    get:
      operationId: listAllEventsAdmin
      tags: [events]
      summary: List every event, optionally filtered by status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          description: Filter to a single status, e.g. pending_review to see the review queue
          schema:
            type: string
            enum: [draft, pending_review, live, cancelled, ended]
      responses:
        "200":
          description: All events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── TICKETS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required:
        - message
...
    TicketTypeSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        currency:
          type: string
        available:
          type: integer
      required:
        - id
        - name
        - price
        - currency
        - available
    EventSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        category:
          type: string
        venue:
          type: string
        city:
          type: string
        country:
          type: string
        countryCode:
          type: string
        eventDate:
          type: string
          format: date-time
        imageUrl:
          type: string
          nullable: true
        imageKey:
          type: string
          nullable: true
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
          nullable: true
        minPrice:
          type: number
        currency:
          type: string
        status:
          type: string
      required:
        - id
        - title
        - category
        - venue
        - city
        - country
        - countryCode
        - eventDate
        - featured
        - minPrice
        - currency
        - status
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/EventSummary"
        - type: object
          properties:
            description:
              type: string
            capacity:
              type: integer
              nullable: true
            latitude:
              type: number
              nullable: true
            longitude:
              type: number
              nullable: true
            ticketTypes:
              type: array
              items:
                $ref: "#/components/schemas/TicketTypeSummary"
            creatorId:
              type: string
          required:
            - description
            - ticketTypes
            - creatorId
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - events
        - total
        - limit
        - offset
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [Music, Art, Food, Heritage, Comedy, Sports, Nightlife]
        venue:
          type: string
        city:
          type: string
        country:
          type: string
        countryCode:
          type: string
        eventDate:
          type: string
          format: date-time
        imageUrl:
          type: string
        capacity:
          type: integer
        tags:
          type: array
          items:
            type: string
        ticketTypes:
          type: array
          items:
            $ref: "#/components/schemas/CreateTicketTypeRequest"
      required:
        - title
        - description
        - category
        - venue
        - city
        - country
        - countryCode
        - eventDate
        - ticketTypes
    CreateTicketTypeRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        totalQuantity:
          type: integer
      required:
        - name
        - price
        - currency
        - totalQuantity

    # ── TICKETS ─────────────────────────────────────────────────────────────
...
    CreatedEventStats:
      type: object
      required: [id, title, category, eventDate, venue, city, status, ticketsSold, revenue, currency]
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        eventDate:
          type: string
          format: date-time
        venue:
          type: string
        city:
          type: string
        status:
          type: string
        ticketsSold:
          type: integer
        revenue:
          type: number
        currency:
          type: string
    CreatorAnalytics:
      type: object
      required: [events, totalRevenue, totalTicketsSold, liveEvents, weeklySales, salesByCity]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/CreatedEventStats"
        totalRevenue:
          type: number
        totalTicketsSold:
          type: integer
        liveEvents:
          type: integer
        weeklySales:
          type: array
          description: Real ticket sales for the last 8 weeks, bucketed by purchase date. Oldest first.
          items:
            $ref: "#/components/schemas/WeeklySalesPoint"
        salesByCity:
          type: array
          description: Real ticket sales grouped by event city, highest first.
          items:
            $ref: "#/components/schemas/CitySalesPoint"
    WeeklySalesPoint:
      type: object
      required: [weekStart, ticketsSold]
      properties:
        weekStart:
          type: string
          format: date-time
        ticketsSold:
          type: integer
    CitySalesPoint:
      type: object
      required: [city, ticketsSold]
      properties:
        city:
          type: string
        ticketsSold:
          type: integer
    UpdateEventStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [draft, pending_review, live, cancelled, ended]
    EventStatusResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [draft, pending_review, live, cancelled, ended]

    # ── GAMIFICATION ──────────────────────────────────────────────────────────
...